requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
casdoor==1.25.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
Django==5.1.2
django-auto-logout==0.5.1
django-cors-headers==4.6.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
frozenlist==1.5.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
mysqlclient==2.2.5
openapi-codec==1.3.2
pycparser==2.22
PyJWT==2.8.0
PyYAML==6.0.2
requests==2.31.0
setuptools==68.2.2
simplejson==3.19.3
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.9.11
